style(backend): apply some suggestions from clippy, improve code style of backend (#1647)

+++ src-tauri/src/launch/commands.rs
@@ -355,11 +355,11 @@ pub fn cancel_launch_process(
   let mut launching_queue = launching_queue_state.lock()?;
 
   // kill process if pid exists
-  if let Some(launching) = launching_queue.last_mut() {
-    if launching.pid != 0 {
-      launching.current_step = 0; // mark as manually cancelled to avoid game error window popping up
-      kill_process(launching.pid)?;
-    }
+  if let Some(launching) = launching_queue.last_mut()
+    && launching.pid != 0
+  {
+    launching.current_step = 0; // mark as manually cancelled to avoid game error window popping up
+    kill_process(launching.pid)?;
   }
 
   Ok(())
